(MAINT) Re-generalize package template
Prior to this commit, the package template was committed with
the version, download URL, and checksum filled out for v0.0.9
instead of using the template values. This caused the script
to always atempt to rerelease 0.0.9 instead of any newer
version released.

This commit returns the values to their template settings to
enable build and automation.